stat collection change for component based ships

+++ engine/src/main/java/de/instinct/engine/stats/model/unit/component/types/CoreStatistic.java
@@ -13,6 +13,6 @@ public class CoreStatistic extends ComponentStatistic {
 	private int timesBuilt;
 	private int timesDestroyed;
 	private int cpUsed;
-	private int resourcesUsed;
+	private float resourcesUsed;
 
 }
